Audit pass: ADRs tracked, doc casing, python tools split, conventions
DevOps / quality / SecOps audit of every change since 3ddaf59d. Full
manifest at /tmp/nest-audit/audit-2026-04-28.md.

Findings + fixes:

A. Untracked ADRs (CRITICAL)
   .gitignore had 'kdb/' and 'adr/' which silently dropped every
   architectural decision committed there. The 5 existing ADRs
   (0001-0005) were never tracked. Removed kdb/ from .gitignore;
   committed those 5 + 6 new ones (0006-0011) covering this session's
   decisions:
     0001 amended: encodings 1/2/3 now active (see 0006)
     0005 superseded by 0010 (search-text exists now)
     0006 Section encodings 1/2/3 ship within v1
     0007 HNSW (0x07) and BM25 (0x08) as optional sections
     0008 model_hash is a granular reproducible fingerprint
     0009 SIMD runtime dispatch (AVX2 / NEON / scalar)
     0010 nest search-text supersedes ADR 0005
     0011 File hygiene: 300 lines per source file

B. Doc casing
   doc/ARCHITECTURE.MD (uppercase ext!), CHANGELOG.md, USAGE.md
   renamed to lowercase kebab: architecture.md, changelog.md,
   usage.md. Top-level UPPERCASE.md kept (GitHub convention).

C. File hygiene >300 lines (session-modified)
   python/tools/measure_presets.py 409 → 277
     extracted _baseline_decoder.py (137) and _bench_runner.py (77)
   python/tools/nest_build_corpus.py 352 → 210
     extracted _corpus_sources.py (169) — 7 dataset loaders + SOURCES
   No first-party Rust or Python source file > 300 lines remains.

D. Convention files
   clippy.toml added (MSRV=1.85, cognitive-complexity-threshold=15,
   too-many-arguments-threshold=7).
   .editorconfig + rustfmt.toml + pyproject.toml verified pre-existing.
   .gitignore: + .ruff_cache/

E. SecOps scan (clean across the board)
   no AWS / GH / Bearer tokens, no hardcoded passwords, no absolute
   /Users or /home paths, no println!/dbg! in lib code, no TODOs.

F. Baseline refresh
   data/measure/baseline.json regenerated at 100 queries (was 50).
   The 50-query baseline had a lucky exact.p95=3.17 that the 1.5×
   headroom couldn't sustain across runs; 100-query at p95=4.44 is
   representative.

G. Cleanup
   /tmp/recall_sweep, /tmp/measure_*.log, /tmp/post_phase*.json
   removed.

Final integrity: ./scripts/release_check.sh — ALL GATES PASS.

  preset       size_MB  size_ratio  recall@10  p50_ms  p95_ms
  exact         119.94      1.000     1.0000    3.48    4.04
  compressed     41.99      0.350     1.0000    3.58    4.49
  tiny           33.93      0.283     0.9920    1.77    3.67
  hybrid         80.17      0.668     1.0000    4.72    6.44
