+++ data/measure/baseline.json
@@ -1,26 +1,25 @@
-baseline: /Users/replicant/dev/nest/data/corpus_next.v1.nest  n=30725 dim=384 dtype=float32
 {
-  "baseline_content_hash": "sha256:b1aac2268deddb2c3d503d4e0e75bf596277c9809671399e6b636a454f6e1ae8",
-  "baseline_file_hash": "sha256:b61b03dd95d9a7c387c2abeaa60e1f4f4ec76ec4a81c5f7d0dc73dbe2050371e",
+  "baseline_content_hash": "sha256:1aa9ea8a0df9cbd9de26fcbe10cc0a4e036f9f4cfac8834f11bded43ec3ef413",
+  "baseline_file_hash": "sha256:d9ec433a868f9a45006756c862c4d6353103128cc1988691d3dd7d3fafbe4324",
   "baseline_path": "/Users/replicant/dev/nest/data/corpus_next.v1.nest",
   "embedding_dim": 384,
   "k": 10,
   "n_chunks": 30725,
-  "n_queries": 50,
+  "n_queries": 100,
   "presets": [
     {
       "build_s": null,
-      "content_hash": "sha256:b1aac2268deddb2c3d503d4e0e75bf596277c9809671399e6b636a454f6e1ae8",
+      "content_hash": "sha256:1aa9ea8a0df9cbd9de26fcbe10cc0a4e036f9f4cfac8834f11bded43ec3ef413",
       "drift_max": 0.0,
       "drift_mean": 0.0,
       "dtype": "float32",
-      "file_hash": "sha256:b61b03dd95d9a7c387c2abeaa60e1f4f4ec76ec4a81c5f7d0dc73dbe2050371e",
+      "file_hash": "sha256:d9ec433a868f9a45006756c862c4d6353103128cc1988691d3dd7d3fafbe4324",
       "has_ann": false,
       "has_bm25": false,
       "is_baseline": true,
-      "p50_ms": 3.072,
-      "p95_ms": 3.1698,
-      "p99_ms": 3.3951,
+      "p50_ms": 3.7198,
+      "p95_ms": 4.4389,
+      "p99_ms": 4.9741,
       "preset": "exact",
       "recall_at_k": 1.0,
       "search_mode": "exact",
@@ -28,18 +27,18 @@ baseline: /Users/replicant/dev/nest/data/corpus_next.v1.nest  n=30725 dim=384 dt
       "size_ratio": 1.0
     },
     {
-      "build_s": 26.48,
-      "content_hash": "sha256:cc6d3247973794b68eef29a89a4a2b3dcb65ef28b360a8aee614843218fcda5e",
+      "build_s": 44.03,
+      "content_hash": "sha256:26f1d985836afd772da7caa107c3da043791a37a0c1ca90c86b6bc32e2a6fdb3",
       "drift_max": 4.9e-05,
       "drift_mean": 1.5e-05,
       "dtype": "float16",
-      "file_hash": "sha256:2b3e8fd2484e154ace5f06e13906d7aa48e05885c7dd1053e73c09a8c106479b",
+      "file_hash": "sha256:8d0f8bb9a19edd69ea626ca85a7f60a9276ff4a790d850467e13d60e42535e3d",
       "has_ann": false,
       "has_bm25": false,
       "is_baseline": false,
-      "p50_ms": 3.3081,
-      "p95_ms": 3.468,
-      "p99_ms": 3.54,
+      "p50_ms": 4.6809,
+      "p95_ms": 15.8572,
+      "p99_ms": 22.3131,
       "preset": "compressed",
       "recall_at_k": 1.0,
       "search_mode": "exact",
@@ -47,42 +46,42 @@ baseline: /Users/replicant/dev/nest/data/corpus_next.v1.nest  n=30725 dim=384 dt
       "size_ratio": 0.3501
     },
     {
-      "build_s": 48.21,
-      "content_hash": "sha256:ba381fe43d9cd3516c9c737948fa494ba49bf5214e5567558bde4eb5af61649c",
-      "drift_max": 0.29369,
-      "drift_mean": 0.006188,
+      "build_s": 96.58,
+      "content_hash": "sha256:1ae7bb77da8ecd3af05232aa2bdcef502d5a20d454fba4a79e6d118ab29d6eb0",
+      "drift_max": 0.001531,
+      "drift_mean": 0.000296,
       "dtype": "int8",
-      "file_hash": "sha256:adf2194d55e9b6e7709b565787ae7e8bc8658cf5ac19bd31756035df5b31e556",
+      "file_hash": "sha256:b3cb225c425dee2fa3f1758348c17a642e85ae37a0b7192884c3986a66e22db9",
       "has_ann": true,
       "has_bm25": false,
       "is_baseline": false,
-      "p50_ms": 0.525,
-      "p95_ms": 0.8211,
-      "p99_ms": 1.1399,
+      "p50_ms": 1.858,
+      "p95_ms": 3.3147,
+      "p99_ms": 3.7048,
       "preset": "tiny",
-      "recall_at_k": 0.97,
+      "recall_at_k": 0.992,
       "search_mode": "ann",
-      "size_mb": 34.6379,
-      "size_ratio": 0.2888
+      "size_mb": 33.9321,
+      "size_ratio": 0.2829
     },
     {
-      "build_s": 68.66,
-      "content_hash": "sha256:4c6e11ad2a8d42580a0c1cb37c03a617ea18cc09f94f03c570df30d7d48c1846",
+      "build_s": 124.22,
+      "content_hash": "sha256:b8c77d7aadb1bb73ec12d621559e30f4eaf83334e0b14ac72afc186659708bab",
       "drift_max": 0.0,
       "drift_mean": 0.0,
       "dtype": "float32",
-      "file_hash": "sha256:5e6a2331c6532c620f2313c15a60fa1694aa926b9018186626248073ff2eeb16",
+      "file_hash": "sha256:1096efb2bbf1a9039e7523d74db86a996fdd09150a669b472e245cea2c53e89e",
       "has_ann": true,
       "has_bm25": true,
       "is_baseline": false,
-      "p50_ms": 4.2858,
-      "p95_ms": 7.0071,
-      "p99_ms": 8.0359,
+      "p50_ms": 5.8,
+      "p95_ms": 9.1069,
+      "p99_ms": 10.3292,
       "preset": "hybrid",
-      "recall_at_k": 0.994,
+      "recall_at_k": 1.0,
       "search_mode": "hybrid",
-      "size_mb": 80.8767,
-      "size_ratio": 0.6743
+      "size_mb": 80.1709,
+      "size_ratio": 0.6684
     }
   ],
   "schema_version": 1,
