+++ doc/architecture.md
