+++ kdb/adr/template.md
@@ -1,0 +1,36 @@
+# ADR NNNN: <title>
+
+- **Status:** Proposed | Accepted | Superseded by ADR-####
+- **Date:** YYYY-MM-DD
+- **Deciders:** <handles>
+
+## Context
+
+Why are we deciding this now? What forces are at play?
+
+## Decision
+
+What did we decide? State the decision in present tense.
+
+## Consequences
+
+### Positive
+
+- ...
+
+### Negative
+
+- ...
+
+### Trade-offs
+
+- What did we explicitly NOT do, and why?
+
+## Alternatives considered
+
+- **Alt A:** rejected because ...
+- **Alt B:** rejected because ...
+
+## References
+
+- ...
